# openapi.yaml
openapi: 3.1.0
info:
  title: scheduler
  version: "1.0.0"
  description: >-
    Cron and one-shot job scheduler: signed HTTP calls to your services, retries with backoff,
    run history, timezone-aware schedules. Part of the atc-web platform; every route under `/v1`
    requires a Bearer API key with a `read` or `write` role (`readwrite` satisfies either).

    Role checks run in Fastify `preValidation`, which executes *before* the request body is
    validated against its JSON Schema. A write request sent with a `read`-only (or missing)
    key therefore always gets `403 FORBIDDEN`, even when the body itself is malformed — the
    body is never parsed against its schema in that case. Only a request presenting a key with
    the required role reaches body validation and can receive `400 VALIDATION_FAILED`.
...
security:
  - bearerAuth: []
paths:
...
  /v1/jobs:
    post:
      operationId: scheduler.jobs.create
      summary: Create a job
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "201":
          description: Job created.
          headers:
            location:
              description: "`/v1/jobs/{name}` of the created job."
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [job]
                properties:
                  job:
                    $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/JobExists"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"

    get:
      operationId: scheduler.jobs.list
      summary: List jobs
      description: Sorted by name; keyset-paginated on name.
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/JobsQ"
        - $ref: "#/components/parameters/JobsTag"
        - $ref: "#/components/parameters/JobsEnabled"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/JobsCursor"
      responses:
        "200":
          description: Page of jobs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/jobs/{name}:
    get:
      operationId: scheduler.jobs.get
      summary: Get a job
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/JobName"
      responses:
        "200":
          description: The job.
          content:
            application/json:
              schema:
                type: object
                required: [job]
                properties:
                  job:
                    $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/JobNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

    patch:
      operationId: scheduler.jobs.patch
      summary: Partially update a job
      description: >-
        Changing `schedule` or setting `enabled: true` on a paused job recomputes the next
        firing; setting `enabled: false` clears it. At least one property is required.
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/JobName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPatch"
      responses:
        "200":
          description: The updated job.
          content:
            application/json:
              schema:
                type: object
                required: [job]
                properties:
                  job:
                    $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/JobNotFound"
        "413":
          $ref: "#/components/responses/BodyTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"

    delete:
      operationId: scheduler.jobs.delete
      summary: Delete a job
      description: Deletes the job and, by database cascade, all of its runs.
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/JobName"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/JobNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/jobs/{name}/run:
    post:
      operationId: scheduler.jobs.run
      summary: Trigger a job manually
      description: >-
        Queues a new `pending` run immediately, regardless of the job's schedule or `enabled`
        bit. Refused with `409 RUN_ACTIVE` when the job already has a `pending`, `running` or
        `retrying` run — only one run may be in flight per job at a time.
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/JobName"
      responses:
        "202":
          description: Run queued.
          content:
            application/json:
              schema:
                type: object
                required: [run]
                properties:
                  run:
                    $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/JobNotFound"
        "409":
          $ref: "#/components/responses/RunActive"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/jobs/{name}/runs:
    get:
      operationId: scheduler.jobs.listRuns
      summary: List a job's runs
      description: >-
        Newest first, keyset-paginated by run id. `404 JOB_NOT_FOUND` when the job does not
        exist (checked before listing). The query schema also accepts a `job` filter (shared
        with `GET /v1/runs`'s schema), but any value supplied there has no effect: this route
        always filters by the `{name}` path parameter.
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/JobName"
        - $ref: "#/components/parameters/RunsStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/RunsBefore"
      responses:
        "200":
          description: Page of runs for this job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/JobNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/runs:
    get:
      operationId: scheduler.runs.list
      summary: List runs across all jobs
      description: Newest first, keyset-paginated by run id.
      tags: [Runs]
      parameters:
        - $ref: "#/components/parameters/RunsJob"
        - $ref: "#/components/parameters/RunsStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/RunsBefore"
      responses:
        "200":
          description: Page of runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/runs/{id}:
    get:
      operationId: scheduler.runs.get
      summary: Get a run
      tags: [Runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: The run.
          content:
            application/json:
              schema:
                type: object
                required: [run]
                properties:
                  run:
                    $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RunNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/runs/{id}/cancel:
    post:
      operationId: scheduler.runs.cancel
      summary: Cancel a queued or retrying run
      description: >-
        Only `pending` and `retrying` runs can be cancelled. A `running` run (an attempt
        currently in flight) cannot be interrupted through this or any other HTTP route —
        cancelling it returns `409 RUN_NOT_CANCELLABLE`, as does cancelling a run already in a
        terminal state (`succeeded`, `failed`, `skipped`, `cancelled`).
      tags: [Runs]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: The cancelled run.
          content:
            application/json:
              schema:
                type: object
                required: [run]
                properties:
                  run:
                    $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RunNotFound"
        "409":
          $ref: "#/components/responses/RunNotCancellable"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/schedule/preview:
    get:
      operationId: scheduler.schedule.preview
      summary: Preview a cron schedule's upcoming firings
      description: >-
        Cron schedules only — there is no equivalent preview for a one-shot (`at`) schedule,
        since a one-shot schedule has exactly one firing, itself.
      tags: [Schedule]
      parameters:
        - name: cron
          in: query
          required: true
          description: Five-field cron expression (or `@hourly`/`@daily`/`@weekly`/`@monthly`/`@yearly`/`@annually`/`@midnight` alias).
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: timezone
          in: query
          required: false
          description: IANA timezone name. Defaults to the server's `DEFAULT_TIMEZONE`.
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - name: count
          in: query
          required: false
          description: Number of upcoming firings to return, 1-50. Defaults to 5.
          schema:
            type: string
            pattern: "^([1-9]|[1-4][0-9]|50)$"
      responses:
        "200":
          description: Normalized cron/timezone and the next firings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewResponse"
        "400":
          $ref: "#/components/responses/InvalidSchedule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/target-keys:
    get:
      operationId: scheduler.targetKeys.list
      summary: List configured target key names
      description: >-
        Names only (sorted), from `TARGET_KEYS` — never the secrets themselves. A job references
        one of these names via `targetKey` to have its outbound call carry that key's bearer
        token, instead of putting `Authorization` in `target.headers` (which is rejected).
      tags: [Reference]
      responses:
        "200":
          description: Configured target key names.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TargetKeysResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/timezones:
    get:
      operationId: scheduler.timezones.list
      summary: List supported IANA timezone names
      description: The full set of `Intl.supportedValuesOf('timeZone')` for this Node runtime.
      tags: [Reference]
      responses:
        "200":
          description: Supported timezone names.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimezonesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/stats:
    get:
      operationId: scheduler.stats.get
      summary: Job and run statistics
      description: >-
        `runs.last24h`/`avgDurationMs24h`/`topFailures24h` cover a fixed trailing 24-hour window.
        `worker.sinceStart` is `null` from an API-only process (no in-process `Worker` to read
        since-start counters from); `worker.inFlight` and `worker.running` are always real and
        durable either way (from the `runs` table / `worker_heartbeat`, not process memory).
      tags: [Reference]
      responses:
        "200":
          description: Aggregate statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <secret>`. Every configured key is compared in constant time; a
        key carries a role (`read`, `write` or `readwrite`) that gates which routes it may call.
        Missing or unrecognized key: `401 UNAUTHORIZED` with a `www-authenticate: Bearer` header.
        Recognized key lacking the required role: `403 FORBIDDEN`.
  parameters:
...
    JobsQ:
      name: q
      in: query
      required: false
      description: Case-sensitive substring match against name and description.
      schema:
        type: string
        minLength: 1
        maxLength: 120
    JobsTag:
      name: tag
      in: query
      required: false
      description: Exact match against one of the job's tags.
      schema:
        type: string
        minLength: 1
        maxLength: 40
    JobsEnabled:
      name: enabled
      in: query
      required: false
      schema:
        type: string
        enum: ["true", "false"]
...
    Limit:
      name: limit
      in: query
      required: false
      description: Page size, 1-200. Defaults to 50.
      schema:
        type: string
        pattern: "^([1-9]|[1-9][0-9]|1[0-9][0-9]|200)$"
...
  responses:
    Unauthorized:
      description: Missing or unrecognized API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: UNAUTHORIZED, message: "missing or invalid API key" }
    Forbidden:
      description: The API key does not have the role required for this route.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: FORBIDDEN, message: "this API key has no write access" }
    RateLimited:
      description: This service's own per-key rate limit was exceeded (distinct from any job-level HTTP outcome).
      headers:
        retry-after:
          schema:
            type: integer
        x-ratelimit-limit:
          schema:
            type: integer
        x-ratelimit-remaining:
          schema:
            type: integer
        x-ratelimit-reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: RATE_LIMITED, message: "rate limit exceeded, retry in 42" }
    BadRequest:
      description: >-
        Either a body/query that fails JSON Schema validation (`VALIDATION_FAILED`) or one that
        passes schema validation but fails a domain rule — `INVALID_SCHEDULE`, `INVALID_TARGET`
        or `UNKNOWN_TARGET_KEY`. See each schema's description for exactly which rule produced
        which code.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            validationFailed:
              value:
                error:
                  code: VALIDATION_FAILED
                  message: "body/schedule must have required property 'cron'"
                  details: [{ path: "/schedule", message: "must have required property 'cron'", params: {} }]
            invalidSchedule:
              value:
                error: { code: INVALID_SCHEDULE, message: "invalid cron expression: expected 5 fields, got 4" }
            invalidTarget:
              value:
                error: { code: INVALID_TARGET, message: "scheme \"ftp:\" is not allowed" }
            unknownTargetKey:
              value:
                error: { code: UNKNOWN_TARGET_KEY, message: "target key \"billing\" is not configured (TARGET_KEYS)" }
...
    BodyTooLarge:
      description: "`target.body`, JSON-encoded, exceeds `MAX_BODY_BYTES`."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: BODY_TOO_LARGE, message: "body is 20000 bytes, limit 16384" }
...
    JobExists:
      description: A job with this name already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: { code: JOB_EXISTS, message: "job \"nightly-export\" already exists" }
...
  schemas:
...
    JobCreate:
      type: object
      additionalProperties: false
      required: [name, schedule, target]
      properties:
        name:
          $ref: "#/components/schemas/JobNameValue"
        description:
          type: string
          maxLength: 500
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
          description: Normalized on save (trimmed, lower-cased, de-duplicated, sorted).
        enabled:
          type: boolean
          description: Defaults to `true`. A disabled job has no `nextRunAt` and never fires on its own.
        schedule:
          $ref: "#/components/schemas/Schedule"
        target:
          $ref: "#/components/schemas/Target"
        targetKey:
          type: [string, "null"]
          maxLength: 64
          description: >-
            Name of a configured `TARGET_KEYS` entry whose bearer token is sent with the
            outbound call. Must name a configured key, else `400 UNKNOWN_TARGET_KEY`.
        timeoutMs:
          type: integer
          minimum: 1
          maximum: 3600000
          description: Defaults to the server's `DEFAULT_TIMEOUT_MS`. Must be between 1000 and the server's `MAX_TIMEOUT_MS`, else `INVALID_TARGET`.
        retry:
          $ref: "#/components/schemas/RetryPolicy"
...
    Job:
      type: object
      required:
        - name
        - description
        - tags
        - enabled
        - schedule
        - target
        - targetKey
        - timeoutMs
        - retry
        - nextRunAt
        - lastRunAt
        - lastStatus
        - createdBy
        - createdAt
        - updatedAt
      properties:
        name:
          $ref: "#/components/schemas/JobNameValue"
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        schedule:
          $ref: "#/components/schemas/Schedule"
        target:
          $ref: "#/components/schemas/Target"
        targetKey:
          type: [string, "null"]
        timeoutMs:
          type: integer
        retry:
          $ref: "#/components/schemas/RetryPolicy"
        nextRunAt:
          type: [string, "null"]
          format: date-time
          description: "`null` when the job is disabled or, for an `at` schedule, once it has fired."
        lastRunAt:
          type: [string, "null"]
          format: date-time
          description: When the job's most recent firing was recorded (`null` before its first firing).
        lastStatus:
          type: [string, "null"]
          enum: [succeeded, failed, skipped, cancelled, null]
          description: Terminal status of the job's most recent firing. Never `pending`/`running`/`retrying` (those are not "recorded outcomes"). `null` before the job's first firing.
        createdBy:
          type: string
          description: The API key id that created the job.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...